Added Comment Section GUI
GUI is functional but needs alot of work with sizing, orientation and layout. I could not figure out how to fi this problem. Also, it is currently set to "test" app. Need to figure out how to make it so it works for apps when selected.

Ignore the appGUI file in this branch, I messed around with it but could not get it to work.